requirements.txt
pandas>=2.1.0
openpyxl>=3.1.0
xlrd>=2.0.1
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
click>=8.1.0
prophet>=1.1.5
statsmodels>=0.14.0
scipy>=1.11.0
numpy>=1.24.0
tqdm>=4.66.0
requests>=2.31.0

# Dashboard
streamlit>=1.31.0
plotly>=5.18.0
